// packages/lib/shared/services/api/generated/graphql.ts
export type Maybe<T> = T | null
// ...
export type Scalars = {
  ID: { input: string; output: string }
  String: { input: string; output: string }
  Boolean: { input: boolean; output: boolean }
  Int: { input: number; output: number }
  Float: { input: number; output: number }
  AmountHumanReadable: { input: string; output: string }
  BigDecimal: { input: string; output: string }
  BigInt: { input: string; output: string }
  Bytes: { input: string; output: string }
  Date: { input: any; output: any }
  GqlBigNumber: { input: string; output: string }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any }
}

export type GqlBalancePoolAprItem = {
  __typename: 'GqlBalancePoolAprItem'
  apr: GqlPoolAprValue
  id: Scalars['ID']['output']
  subItems?: Maybe<Array<GqlBalancePoolAprSubItem>>
  title: Scalars['String']['output']
}

export type GqlBalancePoolAprSubItem = {
  __typename: 'GqlBalancePoolAprSubItem'
  apr: GqlPoolAprValue
  id: Scalars['ID']['output']
  title: Scalars['String']['output']
}

export enum GqlChain {
  Arbitrum = 'ARBITRUM',
  Avalanche = 'AVALANCHE',
  Base = 'BASE',
  Fantom = 'FANTOM',
  Fraxtal = 'FRAXTAL',
  Gnosis = 'GNOSIS',
  Mainnet = 'MAINNET',
  Mode = 'MODE',
  Optimism = 'OPTIMISM',
  Polygon = 'POLYGON',
  Sepolia = 'SEPOLIA',
  Zkevm = 'ZKEVM',
}
// ...
export type GqlPoolTokenDisplay = {
  __typename: 'GqlPoolTokenDisplay'
  address: Scalars['String']['output']
  id: Scalars['ID']['output']
  name: Scalars['String']['output']
  nestedTokens?: Maybe<Array<GqlPoolTokenDisplay>>
  symbol: Scalars['String']['output']
  weight?: Maybe<Scalars['BigDecimal']['output']>
}
// ...
export type GqlPriceRateProviderData = {
  __typename: 'GqlPriceRateProviderData'
  /** The address of the price rate provider */
  address: Scalars['String']['output']
  /** The factory used to create the price rate provider, if applicable */
  factory?: Maybe<Scalars['String']['output']>
  /** The name of the price rate provider */
  name?: Maybe<Scalars['String']['output']>
  /** The filename of the review of the price rate provider */
  reviewFile?: Maybe<Scalars['String']['output']>
  /** Indicates if the price rate provider has been reviewed */
  reviewed: Scalars['Boolean']['output']
  /** A summary of the price rate provider, usually just says safe or unsafe */
  summary?: Maybe<Scalars['String']['output']>
  /** Upgradeable components of the price rate provider */
  upgradeableComponents?: Maybe<Array<Maybe<GqlPriceRateProviderUpgradeableComponent>>>
  /** Warnings associated with the price rate provider */
  warnings?: Maybe<Array<Scalars['String']['output']>>
}

/** Represents an upgradeable component of a price rate provider */
export type GqlPriceRateProviderUpgradeableComponent = {
  __typename: 'GqlPriceRateProviderUpgradeableComponent'
  /** The entry point / proxy of the upgradeable component */
  entryPoint: Scalars['String']['output']
  /** Indicates if the implementation of the component has been reviewed */
  implementationReviewed: Scalars['String']['output']
}
// ...
export type GqlToken = {
  __typename: 'GqlToken'
  /** The address of the token */
  address: Scalars['String']['output']
  /** The chain of the token */
  chain: GqlChain
  /** The chain ID of the token */
  chainId: Scalars['Int']['output']
  /** The coingecko ID for this token, if present */
  coingeckoId?: Maybe<Scalars['String']['output']>
  /** The number of decimal places for the token */
  decimals: Scalars['Int']['output']
  /** The description of the token */
  description?: Maybe<Scalars['String']['output']>
  /** The Discord URL of the token */
  discordUrl?: Maybe<Scalars['String']['output']>
  /** Whether the token is considered an ERC4626 token. */
  isErc4626: Scalars['Boolean']['output']
  /** The logo URI of the token */
  logoURI?: Maybe<Scalars['String']['output']>
  /** The name of the token */
  name: Scalars['String']['output']
  /** The rate provider data for the token */
  priceRateProviderData?: Maybe<GqlPriceRateProviderData>
  /** The priority of the token, can be used for sorting. */
  priority: Scalars['Int']['output']
  /** The rate provider data for the token */
  rateProviderData?: Maybe<GqlPriceRateProviderData>
  /** The symbol of the token */
  symbol: Scalars['String']['output']
  /** The Telegram URL of the token */
  telegramUrl?: Maybe<Scalars['String']['output']>
  /** Indicates if the token is tradable */
  tradable: Scalars['Boolean']['output']
  /** The Twitter username of the token */
  twitterUsername?: Maybe<Scalars['String']['output']>
  /** The website URL of the token */
  websiteUrl?: Maybe<Scalars['String']['output']>
}